requirements.txt
streamlit
langchain
langchain-groq
langchain-community
huggingface-hub
sentence-transformers
faiss-cpu
python-dotenv
